refactor: move stats handler logic to new file
